just add for dont lose project test

+++ project/database/migrations/2014_10_12_000000_create_users_table.php
@@ -26,7 +26,7 @@ class CreateUsersTable extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->tinyInteger('activation')->default(0)->comment('0 => inactive, 1 => active');
             $table->timestamp('activation_date')->nullable();
-            $table->tinyInteger('user_type')->default(0)->comment('0 => user, 1 => admin');
+            $table->integer('user_type')->default(0)->comment('0 => user, 1 => admin');
             $table->tinyInteger('status')->default(0);
             $table->foreignId('current_team_id')->nullable();
             $table->rememberToken();
